Refactor of commandline and making version adaptable to time
