+++ tests/integration/config/behat.yml
@@ -13,13 +13,11 @@ default:
               - admin
             regular_user_password: 123456
             mailhog_url: http://127.0.0.1:8025/api/v2/messages
+            oc_path: ../../
         - CardDavContext:
             baseUrl: http://localhost:8080
         - CalDavContext:
             baseUrl: http://localhost:8080
-        - CommandLineContext:
-            baseUrl: http://localhost:8080
-            ocPath: ../../
         - AppManagementContext:
     federation:
       paths:
