+++ tests/integration/features/status.feature
@@ -2,7 +2,7 @@ Feature: Status
 
   Scenario: Status.php is correct
       When requesting status.php
-      Then the json responded should match with
+      Then the status.php with versions fixed responded should match with
       """
-      {"installed":true,"maintenance":false,"needsDbUpgrade":false,"version":"10.0.3.3","versionstring":"10.0.3","edition":"Community","productname":"ownCloud"}
+      {"installed":true,"maintenance":false,"needsDbUpgrade":false,"version":"X.X.X.X","versionstring":"X.X.X","edition":"Community","productname":"ownCloud"}
       """
